cart item add, delete view completed

+++ backend/src/main/java/com/wandr/backend/dto/product/UpdateProductDTO.java
@@ -2,14 +2,16 @@ package com.wandr.backend.dto.product;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 
 @Data
 public class UpdateProductDTO {
     private String name;
     private String description;
     private Integer quantity;
-    private Double price;
-    private Double reservation_payment;
+    private BigDecimal price;
+    private BigDecimal reservation_payment;
 //    private String image;
 }
